+++ src/main/resources/application.yml
@@ -16,6 +16,7 @@ spring:
 jwt:
   secret:
     key: 7ZWt7ZW0OTntmZTsnbTtjIXtlZzqta3snYTrhIjrqLjshLjqs4TroZzrgpjslYTqsIDsnpDtm4zrpa3tlZzqsJzrsJzsnpDrpbzrp4zrk6TslrTqsIDsnpA=
+
 #springdoc:
 #  swagger-ui:
 #    path: /swagger-ui.html
